Example of mapping a Scala 2.11 large case class (>22 fields) using Slick.
To my knowledge you can't map large case classes like this directly.
The problem is that, even though Scala 2.11 supports large case classes,
it doesn't support the large tuple or function types that would be needed
for a direct mapping.

This example uses Slick's existing HList mapping to produce a query
that returns large case instances. Even though the code in this file
is not pretty, the code that uses these queries will be fine.